requirements.txt
boto3==1.34.162
PyGithub==2.4.0
GitPython==3.1.43
pyyaml==6.0.2
requests==2.32.3
